+++ massmine.scm
@@ -610,7 +610,8 @@ END
      [(equal? curr-task "wikipedia-search") (wikipedia-search (keywords) (language))]
      [(equal? curr-task "wikipedia-text") (wikipedia-text (keywords) (language))]
      [(equal? curr-task "wikipedia-views") (wikipedia-views (keywords) (date))]
-     [(equal? curr-task "wikipedia-page-links") (wikipedia-page-links (keywords) (language))])]
+     [(equal? curr-task "wikipedia-page-links") (wikipedia-page-links (keywords) (language))]
+     [(equal? curr-task "wikipedia-trends") (wikipedia-trends (date))])]
    [(s-starts-with? "google" curr-task)
     (cond
      [(equal? curr-task "google-trends") (google-trends (date))]